microcloud/service/lxd: Use 'lxd init --preseed' command to init LXD
