+++ microcloud/service/lxd_join.go
@@ -1,12 +1,15 @@
 package service
 
 import (
+	"encoding/pem"
 	"fmt"
 	"net/http"
 
 	"github.com/lxc/lxd/client"
+	"github.com/lxc/lxd/lxd/util"
 	"github.com/lxc/lxd/shared"
 	"github.com/lxc/lxd/shared/api"
+	"github.com/lxc/lxd/shared/logger"
 	"github.com/lxc/lxd/shared/version"
 )
 
@@ -42,4 +45,44 @@ func SetupTrust(serverCert *shared.CertInfo, serverName string, targetAddress st
 	}
 
 	return nil
+}
+
+func (s *LXDService) configFromToken(token string) (*api.ClusterPut, error) {
+	joinToken, err := shared.JoinTokenDecode(token)
+	if err != nil {
+		return nil, fmt.Errorf("Invalid cluster join token: %w", err)
+	}
+	config := &api.ClusterPut{
+		Cluster:       api.Cluster{ServerName: s.name, Enabled: true},
+		ServerAddress: util.CanonicalNetworkAddress(s.address, s.port),
+	}
+
+	// Attempt to find a working cluster member to use for joining by retrieving the
+	// cluster certificate from each address in the join token until we succeed.
+	for _, clusterAddress := range joinToken.Addresses {
+		// Cluster URL
+		config.ClusterAddress = util.CanonicalNetworkAddress(clusterAddress, s.port)
+
+		// Cluster certificate
+		cert, err := shared.GetRemoteCertificate(fmt.Sprintf("https://%s", config.ClusterAddress), version.UserAgent)
+		if err != nil {
+			logger.Warnf("Error connecting to existing cluster member %q: %v\n", clusterAddress, err)
+			continue
+		}
+
+		certDigest := shared.CertFingerprint(cert)
+		if joinToken.Fingerprint != certDigest {
+			return nil, fmt.Errorf("Certificate fingerprint mismatch between join token and cluster member %q", clusterAddress)
+		}
+
+		config.ClusterCertificate = string(pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: cert.Raw}))
+
+		break // We've found a working cluster member.
+	}
+
+	if config.ClusterCertificate == "" {
+		return nil, fmt.Errorf("Unable to connect to any of the cluster members specified in join token")
+	}
+
+	return config, nil
 }
